ai-service/requirements.txt
# AI Service Dependencies
# FastAPI framework and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis[asyncio]>=5.0.0

# Authentication
PyJWT>=2.8.0
pyotp>=2.9.0

# HTTP client
httpx>=0.26.0

# Configuration
pydantic-settings>=2.1.0

# Logging
structlog>=24.1.0

# Form data parsing
python-multipart>=0.0.6

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
tiktoken>=0.6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.19.0
